pyproject.toml
[tool.poetry]
name = "breakbot-v2-2"
version = "1.0.0"
description = "Nextgen Break Bot."
authors = ["Your Name <you@example.com>"]
license = "GPLv3"
readme = "README.md"
packages = [{ include = "breaker" }]

[tool.poetry.dependencies]
python = "^3.12"
nonebot2 = { extras = ["fastapi"], version = "^2.3.1" }
nonebot-adapter-onebot = "^2.4.3"
nonebot-plugin-alconna = "^0.47.0"
colored = "^2.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
